Restructure `api` module
- Move OidcException to root package
- Rename `app.quickcase.spring.oidc.oidc` package to `app.quickcase.spring.oidc.userinfo` to avoid duplicated package name
- Move UserInfo and UserPreferences classes to renamed `userinfo` package to regroup all strongly related classes in dedicated package
