+++ api/src/main/java/app/quickcase/spring/oidc/OidcException.java
@@ -1,4 +1,4 @@
-package app.quickcase.spring.oidc.oidc;
+package app.quickcase.spring.oidc;
 
 import org.springframework.security.core.AuthenticationException;
