+++ api/src/main/java/app/quickcase/spring/oidc/authentication/QuickcaseClientAuthentication.java
@@ -3,7 +3,7 @@ package app.quickcase.spring.oidc.authentication;
 import app.quickcase.spring.oidc.AccessLevel;
 import app.quickcase.spring.oidc.organisation.OrganisationProfile;
 import app.quickcase.spring.oidc.SecurityClassification;
-import app.quickcase.spring.oidc.UserInfo;
+import app.quickcase.spring.oidc.userinfo.UserInfo;
 import org.springframework.security.core.GrantedAuthority;
 
 import java.util.Collection;
