+++ api/src/main/java/app/quickcase/spring/oidc/authentication/QuickcaseUserAuthentication.java
@@ -3,7 +3,7 @@ package app.quickcase.spring.oidc.authentication;
 import app.quickcase.spring.oidc.AccessLevel;
 import app.quickcase.spring.oidc.organisation.OrganisationProfile;
 import app.quickcase.spring.oidc.SecurityClassification;
-import app.quickcase.spring.oidc.UserInfo;
+import app.quickcase.spring.oidc.userinfo.UserInfo;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.GrantedAuthority;
