+++ api/src/main/java/app/quickcase/spring/oidc/userinfo/DefaultUserInfoGateway.java
@@ -1,4 +1,4 @@
-package app.quickcase.spring.oidc.oidc;
+package app.quickcase.spring.oidc.userinfo;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
