+++ api/src/main/java/app/quickcase/spring/oidc/userinfo/UserInfoService.java
@@ -1,6 +1,4 @@
-package app.quickcase.spring.oidc.oidc;
-
-import app.quickcase.spring.oidc.UserInfo;
+package app.quickcase.spring.oidc.userinfo;
 
 public interface UserInfoService {
     UserInfo loadUserInfo(String expectedSubject, String accessToken);
