+++ api/src/test/java/app/quickcase/spring/oidc/authentication/QuickcaseAuthenticationConverterTest.java
@@ -1,8 +1,8 @@
 package app.quickcase.spring.oidc.authentication;
 
 import app.quickcase.spring.oidc.organisation.OrganisationProfile;
-import app.quickcase.spring.oidc.UserInfo;
-import app.quickcase.spring.oidc.oidc.UserInfoService;
+import app.quickcase.spring.oidc.userinfo.UserInfo;
+import app.quickcase.spring.oidc.userinfo.UserInfoService;
 import app.quickcase.spring.oidc.AccessLevel;
 import app.quickcase.spring.oidc.SecurityClassification;
 import org.hamcrest.Matchers;
