+++ keycloak/src/main/java/app/quickcase/spring/oidc/QuickcaseSecurityConfig.java
@@ -3,11 +3,11 @@ package app.quickcase.spring.oidc;
 import app.quickcase.spring.oidc.authentication.QuickcaseAuthenticationConverter;
 import app.quickcase.spring.oidc.keycloak.KeycloakQuickcaseSecurityDsl;
 import app.quickcase.spring.oidc.keycloak.oidc.KeycloakUserInfoExtractor;
-import app.quickcase.spring.oidc.oidc.DefaultUserInfoGateway;
-import app.quickcase.spring.oidc.oidc.DefaultUserInfoService;
-import app.quickcase.spring.oidc.oidc.UserInfoExtractor;
-import app.quickcase.spring.oidc.oidc.UserInfoGateway;
-import app.quickcase.spring.oidc.oidc.UserInfoService;
+import app.quickcase.spring.oidc.userinfo.DefaultUserInfoGateway;
+import app.quickcase.spring.oidc.userinfo.DefaultUserInfoService;
+import app.quickcase.spring.oidc.userinfo.UserInfoExtractor;
+import app.quickcase.spring.oidc.userinfo.UserInfoGateway;
+import app.quickcase.spring.oidc.userinfo.UserInfoService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
